Fix comunity supported check
