+++ frontend/packages/kubevirt-plugin/locales/en/kubevirt-plugin.json
@@ -689,8 +689,6 @@
   "You can exit this screen and return later by launching console from the following virtual machine.": "You can exit this screen and return later by launching console from the following virtual machine.",
   "This flow will require you to generalize your custom boot source.": "This flow will require you to generalize your custom boot source.",
   "Learn how to generalize a boot source": "Learn how to generalize a boot source",
-  "Community Supported": "Community Supported",
-  "Community": "Community",
   "Create new Template": "Create new Template",
   "Create a new customized template based on this template": "Create a new customized template based on this template",
   "Delete Template": "Delete Template",
@@ -739,6 +737,8 @@
   "This operating system boot source was added to the cluster by user on {{date}}": "This operating system boot source was added to the cluster by user on {{date}}",
   "Unknown source": "Unknown source",
   "Create": "Create",
+  "Community Supported": "Community Supported",
+  "Community": "Community",
   "See template details for support.": "See template details for support.",
   "Supported operating systems are labeled below.": "Supported operating systems are labeled below.",
   "Learn more about Red Hat support": "Learn more about Red Hat support",
