+++ frontend/packages/kubevirt-plugin/src/components/create-vm/tabs/select-template.scss
@@ -43,7 +43,3 @@
 .kv-select-template__tile-desc {
   --pf-l-stack--m-gutter--MarginBottom: var(--pf-global--spacer--sm);
 }
-
-.kv-select-template__support-label {
-  text-align: right;
-}
